Title: 특정 문자로 시작하는 문자열, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 250918/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.cpp
@@ -24,6 +24,6 @@ int main() {
         }
     }
     
-    cout << cnt << " " << fixed << setprecision(2) << double(sum)/double(cnt);
+    cout << cnt << " " << fixed << setprecision(2) << double(sum)/cnt;
     return 0;
 }
